coqui-rewrite: More refactoring
